--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "vendor/iceberg-python"]
-	path = vendor/iceberg-python
-	url = git@github.com:apache/iceberg-python.git
