+++ pyproject.toml
@@ -1,0 +1,36 @@
+[tool.poetry]
+name = "iceberg-rest"
+version = "0.1.0"
+description = "An Apache Iceberg REST Catalog implementation in Python."
+authors = ["Kevin Liu <kevinjqliu@users.noreply.github.com>", "GoldToothRichards <crabtr26@proton.me>"]
+readme = "README.md"
+packages = [{include = "iceberg_rest", from = "src"}]
+
+[tool.poetry.dependencies]
+python = "^3.10"
+fastapi = "^0.111.0"
+psycopg2-binary = "^2.9.9"
+pyiceberg = {git = "https://github.com/apache/iceberg-python"}
+s3fs = "^2024.6.0"
+sqlalchemy = "^2.0.30"
+uvicorn = "^0.30.1"
+
+[tool.poetry.group.dev]
+optional = true
+
+[tool.poetry.group.dev.dependencies]
+moto = "^5.0.9"
+pandas = "^2.2.2"
+pre-commit = "^3.7.1"
+pyarrow = "^16.1.0"
+pyright = "^1.1.366"
+pyspark = "^3.5.1"
+pytest = "7.4.4"
+pytest-icdiff = "^0.9"
+pytest-lazy-fixture = "0.6.3"
+ruff = "^0.4.8"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
