--- requirements.txt
@@ -1,13 +1,0 @@
-fastapi
-# install pyiceberg from vendor directory. This contains temporary fixes
-./vendor/iceberg-python                                             
-sqlalchemy
-pyiceberg[pyarrow]
-pyiceberg[s3fs]
-ruff
-# tests
-pytest==7.4.4
-pytest-lazy-fixture==0.6.3
-moto
-pyspark
-pandas
