+++ src/iceberg_rest/exception.py
